Tutorial 4 work upload
Changes made to website within this upload:
I developed 1 new page: "Resume and Professional Profile". In this page I put a slightly blurred image of my resume with instructions next to it to click on the image to download my resume. For this I had to upload the image of my resume and then link my pdf to the image so people can click on it to download it. Underneath my resume I also added my "Professional Profile" which I developed to be sort of like a general cover letter. It pairs with my resume to give people a bit of insight into me (even though I have the "about me" page. This is slightly more formal than that page).
I changed the layout of my "Unit Plans and Resource Samples" page to fit the requirements of the tutorial this week. Before these changes it was more spread out with a description of the lesson on the top and an image of each page of the resource beneath that text. As per the tutorial work, I changed this page into a grid layout. This condensed the information on this page slightly and meant it would be tricky to have the images of the resources as I had them before. Because of this I changed it so that the image was still of the pages but it was now all of the resource pages in a pile (sort of like a little preview of the first page and a tiny bit of the other pages behind), and then if you click on that image it will download the whole resource for the person to use. I much prefer this layout, I think it looks cleaner and a lot more formal than my old clunky design.
The tutorial asked me to create a contact form on my "Contact Me" page. I already had this in previous page iterations, but in this tutorial I developed it further, slightly changing the code so it was a bit neater (by removing almost all of the divs and instead using fieldset and legend), and also adding a yellow border around it so it stands out slightly more.
As a part of adding the new page for my resume, I also had to adjust all of the nav menus to include a link to this page. 

*Images again uploaded to GitHub as individual photos, not in the folder. I'm unsure why this happened again as I used the method that worked last week.*