Tutorial 5 work upload: contact.html continued
After testing the contact form using FormSubmit.co, I felt the redirect back to the website was boring (bland/beige colours and text) so I added a "Thank You" page to my website that the form will direct you to after the form sender presses send. I haven't added a link to this page anywhere else on my website so I believe it would only be accessible through the form.
The "Thank You" page doesn't have a nav menu, rather just a link at the bottom to direct them back to the home page. It has a thank you image (inside a yellow border to contain the image a bit better as it was spreading too wide on the page) that follows the theme I have set across my whole website as well as a short message saying I will respond to their contact form asap.
I adjusted the size of the text on this page as there isn't much on the page so there was lots of blank space and it looked too small.

This page is very simple on purpose as I don't want it to be very flashy, just something small to say thank you for sending the form and redirect them back to the website, while also keeping it on theme with what I want my users to experience and see on my website!