11/09/25 edits
Small edit to the worksamples.html page to get access to the thank you page. I added a link to the thank you page in the nav bar so that I can access the actual link to this page as currently the link that I'm using for it doesn't work on devices other than mine as its my server's link. Once I get the actual accessible link for the thank you page I will be removing it from the nav bar on the work samples page as I don't want it to be accessible through my website. It should just be a redirect page.