11/09/25 edits (2)
Used the previous commit to access the live link to the thank you page. Reuploading the work samples page removing the link to the thank you page from the nav bar. Also increased the size of the send button on my contact form as that was very small in relation to everything else around it, and also re-centred/ aligned the email symbol image with my email address as they were misaligned.