Assessment 3 initial upload
Initial upload of website for A3. 
In this upload I put all of the text from my original website in A1 and A2 into the template that I chose for A3. I formatted the text on each of my pages/pop up's and added comment tags throughout where I need to add images or specific formatting. 
This text includes all the headings and text/information on pages as well as the home screen nav menu and the website footer.
I also changed the background of the whole website in this upload so that it fits my theme as well as changed the diamond icon on the home page into the flower image that I put on my website in A2.
I started to edit the contact form, it is now linked to the formsubmit link like my initial website, though I am still working out how to make it so when you click "send message", it redirects to the thank you pop up page.
I also need to include all of my javascript features I chose in A2.
Lastly, I uploaded almost all of the images I use throughout my website into the images folder so that they are ready to be written into my code.