A3 2nd update
In this upload I fixed minor spelling errors throughout the website. I uploaded additional images to the images folder as well as created and uploaded pdfs to the files folder. I adjusted the fade in and out time of the background image when the page loads. I made the table for my major and minor units on my about me page, I also edited the colour of this table. Almost all images were added to their pages, I still need to reposition some of these images as well as add the hyperlinks to these images so that they will download their respective document when clicked.