A3 3rd update (3/3)
It's saying this upload is too big so I'm splitting it into 3:

In this upload/update I fixed an issue where the hacksaw file images weren't fitting next to each other. I removed bottom border (dotted line) below images with links on them. I made "image-with-text" to align image to right of texts where needed. I added an email symbol next to my email on the contact page. I also changed the colour of the send message button to fit the design of my A2. I added a profile image to my about me page. To do so I used code from "image-with-text" but altered it so that there was a smaller gap between the text and the image is larger. I also added all pdf links to images where needed.