A3 4th update 3/3
Splitting into 3 uploads as it will be too big to upload in 1.

Completed my 2 JS features in this upload. On my resume page, it now has a pop up message and form appear when you click on my resume thumbnail, this delays the download of my resume until the form is completed. On the contact me page, a pop up message now appears, thanking the person by their name (the one on the form), in a personalised message. Also uploaded PDFs to all CAD images on the university projects page. Removed the thank you page as this became too hard to write into the code. It was too hard to get the formsubmit page to link back to that page rather than the home page.